// src/interfaces/IGameStatisticData.ts
export interface IGameStatisticData {
    ranking: number;
    practicePlayedEasy: number;
    practiceVictoryEasy: number;
    practicePlayedMedium: number;
    practiceVictoryMedium: number;
    practicePlayedHard: number;
    practiceVictoryHard: number;
    onlinePlayed: number;
    onlineVictory: number;
}